[feat] 리뷰페이지 테스트 console.log 제거 , 폴더명 변경
