+++ app/(user)/my/review/review.module.scss
@@ -37,6 +37,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
+  padding-bottom: 5.5rem;
 }
 
 .write_content,
@@ -62,6 +63,7 @@
 }
 
 .product_link {
+  position: relative;
   display: block;
   border-bottom: 1px solid #e3e3e3;
   padding-bottom: 1.5rem;
@@ -95,11 +97,13 @@
 }
 
 .product_eng_name {
+  padding-right: 7rem;
+  color: #222;
+  margin: 0 0 .5rem;
   font-size: 1.1rem;
   font-weight: 600;
-  color: $black;
-  margin: 0 0 0.5rem 0;
   line-height: 1.3;
+  text-wrap: auto;
 }
 
 .product_kor_name {
@@ -115,17 +119,18 @@
   margin: 0 0 1rem 0;
 }
 
+.review_date{
+  position: absolute;
+  top: 0;
+  right: 0rem;
+  display: inline-block;
+  font-size: .9rem;
+}
 .rating_section {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
   gap: 0.5rem;
-}
-
-.rating_text {
-  font-size: 1.1rem;
-  color: #666;
-  margin: 0;
 }
 
 .rating_label{    
@@ -140,4 +145,21 @@
   font-size: 1.1rem;
   font-style: italic;
   margin: 0;
+}
+
+.review_tags {
+  gap: 12px;
+  display: flex
+;
+  flex-wrap: wrap;
+}
+
+.review_tag {
+  text-align: center;
+  color: #333;
+  background-color: #f5f5f5;
+  border: 1px solid #e0e0e0;
+  border-radius: 8px;
+  padding: .5rem;
+  font-size: 1rem;
 }
