+++ backend/reviews/repositories/productRepository.ts
@@ -10,8 +10,6 @@ export interface ProductRepository {
 export class PrismaProductRepository implements ProductRepository {
   async findProductById(productId: number): Promise<Product | null> {
     try {
-      console.log('🔍 상품 조회 시작 - productId:', productId);
-      
       const product = await prisma.product.findUnique({
         where: { id: productId },
         include: {
@@ -20,11 +18,8 @@ export class PrismaProductRepository implements ProductRepository {
       });
       
       if (!product) {
-        console.log('❌ 상품을 찾을 수 없음 - productId:', productId);
         return null;
       }
-      
-      console.log('✅ 상품 조회 성공:', product.engName || product.korName);
       
       return {
         id: product.id,
@@ -38,12 +33,7 @@ export class PrismaProductRepository implements ProductRepository {
         } : undefined
       };
     } catch (error) {
-      console.error('❌ 상품 조회 실패:', error);
-      console.error('❌ 오류 상세:', {
-        name: error instanceof Error ? error.name : 'Unknown',
-        message: error instanceof Error ? error.message : 'Unknown error',
-        stack: error instanceof Error ? error.stack : 'No stack trace'
-      });
+      console.error('상품 조회 실패:', error);
       throw new Error('상품을 조회할 수 없습니다.');
     }
   }
