+++ backend/reviews/repositories/reviewRepository.ts
@@ -6,7 +6,7 @@ import prisma from '../../utils/prisma';
 export interface ReviewRepository {
   findReviewsByUserId(userId: string): Promise<Review[]>;
   findReviewByUserAndProduct(userId: string, productId: number): Promise<Review | null>;
-  findReviewByOrderId(orderId: number): Promise<Review | null>; // orderId로 리뷰 찾기 추가
+  findReviewByOrderId(orderId: number): Promise<Review | null>;
   createReview(review: Omit<Review, 'createdAt'>): Promise<Review>;
 }
 
@@ -28,7 +28,7 @@ export class PrismaReviewRepository implements ReviewRepository {
       return reviews.map(review => ({
         userId: review.userId,
         productId: review.productId,
-        orderId: review.orderId, // orderId 추가
+        orderId: review.orderId,
         contents: review.contents,
         rating: review.rating || 0,
         reviewImages: review.reviewImages || undefined,
@@ -42,12 +42,10 @@ export class PrismaReviewRepository implements ReviewRepository {
 
   async findReviewByUserAndProduct(userId: string, productId: number): Promise<Review | null> {
     try {
-      const review = await prisma.review.findUnique({
+      const review = await prisma.review.findFirst({
         where: {
-          userId_productId: {
-            userId,
-            productId
-          }
+          userId,
+          productId
         }
       });
       
@@ -56,7 +54,7 @@ export class PrismaReviewRepository implements ReviewRepository {
       return {
         userId: review.userId,
         productId: review.productId,
-        orderId: review.orderId, // orderId 추가
+        orderId: review.orderId,
         contents: review.contents,
         rating: review.rating || 0,
         reviewImages: review.reviewImages || undefined,
@@ -93,8 +91,6 @@ export class PrismaReviewRepository implements ReviewRepository {
 
   async createReview(review: Omit<Review, 'createdAt'>): Promise<Review> {
     try {
-      console.log('🔍 리뷰 생성 시도:', review); // 디버깅 로그 추가
-      
       const createdReview = await prisma.review.create({
         data: {
           userId: review.userId,
@@ -102,31 +98,22 @@ export class PrismaReviewRepository implements ReviewRepository {
           contents: review.contents,
           rating: review.rating,
           reviewImages: review.reviewImages,
-          orderId: review.orderId // orderId 필수 필드로 변경
+          orderId: review.orderId
         }
       });
-      
-      console.log('✅ 리뷰 생성 성공:', createdReview); // 성공 로그 추가
       
       return {
         userId: createdReview.userId,
         productId: createdReview.productId,
-        orderId: createdReview.orderId, // orderId 추가
+        orderId: createdReview.orderId,
         contents: createdReview.contents,
         rating: createdReview.rating || 0,
         reviewImages: createdReview.reviewImages || undefined,
         createdAt: createdReview.createdAt || new Date()
       };
     } catch (error) {
-      console.error('❌ 리뷰 생성 실패 - 상세 오류:', error); // 상세 오류 로그
-      console.error('❌ 오류 타입:', error instanceof Error ? error.constructor.name : typeof error);
-      console.error('❌ 오류 메시지:', error instanceof Error ? error.message : 'Unknown error');
-      console.error('❌ 오류 스택:', error instanceof Error ? error.stack : 'No stack trace');
-      
       if (error && typeof error === 'object' && 'code' in error) {
         const prismaError = error as any;
-        console.error('❌ Prisma 오류 코드:', prismaError.code);
-        console.error('❌ Prisma 오류 메타:', prismaError.meta);
         if (prismaError.code === 'P2002') {
           throw new Error('이미 해당 상품에 대한 리뷰를 작성했습니다.');
         }
@@ -139,70 +126,20 @@ export class PrismaReviewRepository implements ReviewRepository {
 
 // Prisma를 사용한 주문 리포지토리 구현
 export class PrismaOrderRepository implements OrderRepository {
-  // 사이즈 정보 추출 헬퍼 함수 - 이제 사용하지 않음
-  // private extractSizeFromProduct(product: any): string {
-  //   try {
-  //     if (!product.productOptionMappings || product.productOptionMappings.length === 0) {
-  //       return '사이즈 정보 없음';
-  //     }
-
-  //     // 첫 번째 옵션 타입에서 사이즈 정보 찾기
-  //     const optionMapping = product.productOptionMappings[0];
-  //     if (optionMapping?.optionType?.optionValue && optionMapping.optionType.optionValue.length > 0) {
-  //       // 사이즈 관련 옵션 값들 (예: S, M, L, 250, 265 등)
-  //       const sizeValues = optionMapping.optionType.optionValue;
-  //       return sizeValues[0].name; // 첫 번째 옵션 값 반환
-  //     }
-
-  //     return '사이즈 정보 없음';
-  //   } catch (error) {
-  //     console.error('사이즈 정보 추출 실패:', error);
-  //     return '사이즈 정보 없음';
-  //   }
-  // }
-
   async findOrdersByUserId(userId: string): Promise<Order[]> {
     try {
-      console.log('🔍 Prisma 쿼리 시작 - userId:', userId);
-      
       const orders = await prisma.order.findMany({
         where: { userId },
         include: {
           product: {
             include: {
               category: true,
-              // productOptionMappings 제거 - 더 이상 필요하지 않음
             },
           },
-          optionValue: true, // optionValue 정보 포함
+          optionValue: true,
         },
         orderBy: { createdAt: 'desc' }
       });
-      
-      console.log('✅ Prisma 쿼리 성공 - 주문 개수:', orders.length);
-      
-      // 첫 번째 주문의 Product 정보 로그
-      if (orders.length > 0 && orders[0].product) {
-        console.log('🔍 첫 번째 주문의 Product 정보:', {
-          id: orders[0].product.id,
-          thumbnailImage: orders[0].product.thumbnailImage,
-          engName: orders[0].product.engName,
-          korName: orders[0].product.korName,
-          category: orders[0].product.category,
-          optionValue: orders[0].optionValue, // optionValue 정보 추가
-        });
-        
-        // optionValue 정보 디버깅 로그 추가
-        if (orders[0].optionValue) {
-          console.log('🔍 OptionValue 정보:', {
-            id: orders[0].optionValue.id,
-            name: orders[0].optionValue.name,
-            typeId: orders[0].optionValue.typeId // any 타입 제거
-          });
-        } else {
-          console.log('❌ OptionValue 정보가 없음');
-        }
-      }
       
       return orders.map(order => ({
         id: order.id,
@@ -225,19 +162,14 @@ export class PrismaOrderRepository implements OrderRepository {
             korName: order.product.category.korName || '카테고리 없음'
           } : undefined
         } : undefined,
-        optionValue: order.optionValue ? {  // optionValue 정보 추가
+        optionValue: order.optionValue ? {
           id: order.optionValue.id,
           name: order.optionValue.name,
           typeId: order.optionValue.typeId
         } : undefined
       }));
     } catch (error) {
-      console.error('❌ 주문 조회 실패:', error);
-      console.error('❌ 오류 상세:', {
-        name: error instanceof Error ? error.name : 'Unknown',
-        message: error instanceof Error ? error.message : 'Unknown error',
-        stack: error instanceof Error ? error.stack : 'No stack trace'
-      });
+      console.error('주문 조회 실패:', error);
       throw new Error('주문을 조회할 수 없습니다.');
     }
   }
@@ -250,10 +182,9 @@ export class PrismaOrderRepository implements OrderRepository {
           product: {
             include: {
               category: true,
-              // productOptionMappings 제거
             },
           },
-          optionValue: true, // optionValue 정보 포함
+          optionValue: true,
         }
       });
       
@@ -280,7 +211,7 @@ export class PrismaOrderRepository implements OrderRepository {
             korName: order.product.category.korName || '카테고리 없음'
           } : undefined
         } : undefined,
-        optionValue: order.optionValue ? {  // optionValue 정보 추가
+        optionValue: order.optionValue ? {
           id: order.optionValue.id,
           name: order.optionValue.name,
           typeId: order.optionValue.typeId
