+++ components/my/ActionGrid/actionGrid.module.scss
@@ -7,4 +7,8 @@
   text-decoration: none;
   color: inherit;
   cursor: pointer;
+  gap: 8px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
