+++ store/userStore.ts
@@ -17,6 +17,7 @@ export interface User {
 interface UserStore {
   // 상태
   user: User | null;
+  reviewCount: number | null; // 사용자의 리뷰 수 추가
   isLoading: boolean;
   error: string | null;
 
@@ -25,11 +26,14 @@ interface UserStore {
   updateUser: (updates: Partial<User>) => void;
   clearUser: () => void;
   fetchUserData: () => Promise<void>;
+  // 리뷰 수만 업데이트하는 액션 추가 (리뷰 작성/삭제 시 사용)
+  updateReviewCount: (count: number) => void;
 }
 
 export const useUserStore = create<UserStore>((set) => ({
   // 초기 상태
   user: null,
+  reviewCount: null, // 리뷰 수 초기값
   isLoading: false,
   error: null,
 
@@ -43,27 +47,52 @@ export const useUserStore = create<UserStore>((set) => ({
     })),
 
   // 사용자 정보 초기화
-  clearUser: () => set({ user: null, error: null }),
+  clearUser: () => set({ user: null, reviewCount: null, error: null }),
 
-  // 서버에서 사용자 정보 가져오기
+  // 리뷰 수 업데이트 (리뷰 작성/삭제 시 사용)
+  updateReviewCount: (count: number) => set({ reviewCount: count }),
+
+  // 서버에서 사용자 정보와 리뷰 수를 함께 가져오기
   fetchUserData: async () => {
     set({ isLoading: true, error: null });
 
     try {
-      const response = await fetch('/api/user/profile');
+      // 1. 사용자 프로필 정보 가져오기
+      const profileResponse = await fetch('/api/user/profile');
 
-      if (!response.ok) {
-        // 더 자세한 에러 정보 로깅
+      if (!profileResponse.ok) {
         console.error('API Response Error:', {
-          status: response.status,
-          statusText: response.statusText,
-          url: response.url
+          status: profileResponse.status,
+          statusText: profileResponse.statusText,
+          url: profileResponse.url
         });
-        throw new Error(`사용자 정보 조회 실패 (${response.status}: ${response.statusText})`);
+        throw new Error(`사용자 정보 조회 실패 (${profileResponse.status}: ${profileResponse.statusText})`);
       }
 
-      const data = await response.json();
-      set({ user: data.user, isLoading: false });
+      const profileData = await profileResponse.json();
+      
+      // 2. 사용자 리뷰 수 가져오기 (사용자 ID가 있을 때만)
+      let reviewCount = 0;
+      if (profileData.user?.id) {
+        try {
+          const reviewResponse = await fetch('/api/reviews/orders');
+          if (reviewResponse.ok) {
+            const reviewData = await reviewResponse.json();
+            reviewCount = reviewData.data?.myReviews?.length || 0;
+          }
+        } catch (reviewError) {
+          console.error('리뷰 수 조회 실패:', reviewError);
+          // 리뷰 수 조회 실패해도 사용자 정보는 설정
+          reviewCount = 0;
+        }
+      }
+
+      // 3. 사용자 정보와 리뷰 수를 함께 설정
+      set({ 
+        user: profileData.user, 
+        reviewCount: reviewCount,
+        isLoading: false 
+      });
     } catch (error) {
       console.error('Fetch User Data Error:', error);
       set({
